Fix issue with Linux mono
Without the dlls in the TargetDir for Assembly-CSharp, the prepatcher
failed to run due to an issue with Cecil. However, what we  had, having
it in the directory of the build, results in a missing method for LINQ.
So I've swapped to having an output directory for the prepatcher, which
fixes both of these problems. It's also a bit cleaner than before
anyways.