Avoid throwing when constructing mods with soft dependencies missing (#94)
* Avoid throwing when constructing certain mods with soft dependencies missing

The particular error this addresses is when some types in the assembly derive from types that are unavailable.

Not throwing here allows soft references to other mods; this will be common with extensions to randomizer, where mods might want to define items to be randomized, but also provide their functionality in a vanilla game.

* Fix previous commit

Because if we define the Modding.Extensions namespace then I have to replace calls to the Extensions class with global::Extensions and it's not worth the effort